Prototype metadata-only filtering
Builds on a prototype augur interface for metadata-only filtering [1] to speed
up the ncov subsampling steps. Specifically, subsampling writes out lists of
selected strain names to text files instead of writing out sequences. Then the
"combine samples" step extracts sequences from the unique list of strain names
with samtools faidx. The resulting workflow requires substantially less I/O and
speeds up subsampling from an average of 163 seconds per job (for 12 jobs in a
Nextstrain global build) to 36 seconds per job.

One downside to the current metadata-only approach in Augur that reveals itself
here is that its possible to have mismatches between metadata and sequences such
that a strain has metadata but does not have a sequence in the filtered
FASTA. One way around this issue is to export the filtered metadata earlier in
the pipeline. For now, we use samtools faidx with the `-c` flag to ignore
missing sequences.

[1] https://github.com/nextstrain/augur/pull/679